Restructured to support playlists WIP
